+++ src/models/qwen3_5/generate.rs
@@ -25,7 +25,7 @@ use crate::{
 pub struct Qwen3_5GenerateModel<'a> {
     chat_template: ChatTemplate<'a>,
     tokenizer: TokenizerModel,
-    pre_processor: Qwen3VLProcessor,
+    pre_processor: Option<Qwen3VLProcessor>,
     qwen3_5: Qwen3_5Model,
     device: Device,
     eos_token_id: u32,
@@ -57,7 +57,7 @@ impl<'a> Qwen3_5GenerateModel<'a> {
         Ok(Self {
             chat_template,
             tokenizer,
-            pre_processor,
+            pre_processor: Some(pre_processor),
             qwen3_5,
             device,
             eos_token_id,
@@ -84,26 +84,26 @@ impl<'a> Qwen3_5GenerateModel<'a> {
         let mut reader = std::fs::File::open(model_file)?;
         let content = gguf_file::Content::read(&mut reader)?;
         let device = get_device(device);
-        let mut gguf = Gguf::new(content, reader, device.clone());
+        let mut model_gguf = Gguf::new(content, reader, device.clone());
 
-        let chat_template_str = gguf
+        let chat_template_str = model_gguf
             .get_matedata("tokenizer.chat_template")?
             .to_string()?
             .clone();
         let chat_template = ChatTemplate::str_init(&chat_template_str)?;
-        let tokenizer = gguf.build_tokenizer(Some(false), Some(false), Some(false))?;
-        let dtype = match gguf.get_matedata("general.type") {
-            Ok(v) => match v.to_u32() {
-                Ok(0) => DType::F32,
-                Ok(1) => DType::F16,
-                _ => DType::F16,
-            },
-            Err(_) => DType::F16,
+        let tokenizer = model_gguf.build_tokenizer(Some(false), Some(false), Some(false))?;
+        let (pre_processor, mut mmproj_gguf) = if let Some(mmproj_f) = mmproj_file {
+            let mut reader = std::fs::File::open(mmproj_f)?;
+            let content = gguf_file::Content::read(&mut reader)?;
+            let mmproj_gguf = Gguf::new(content, reader, device.clone());
+            let processor = Qwen3VLProcessor::new_qwen3_5_default(&device, DType::F32)?;
+            (Some(processor), Some(mmproj_gguf))
+        } else {
+            (None, None)
         };
-        let pre_processor = Qwen3VLProcessor::new_qwen3_5_default(&device, dtype)?;
 
         // let eos_token_id = gguf.get_matedata("tokenizer.ggml.eos_token_id")?.to_u32()?;
-        let qwen3_5 = Qwen3_5Model::new_from_gguf(&mut gguf, &device)?;
+        let qwen3_5 = Qwen3_5Model::new_from_gguf(&mut model_gguf, mmproj_gguf.as_mut(), &device)?;
         let stem = std::path::Path::new(model_file)
             .file_stem() // 获取文件名主干（不含扩展名）
             .and_then(|s| s.to_str())
@@ -131,18 +131,28 @@ impl<'a> GenerateModel for Qwen3_5GenerateModel<'a> {
             get_logit_processor(temperature.into(), top_p.into(), Some(20), seed);
         // let mut logit_processor = get_logit_processor(mes.temperature, mes.top_p, None, seed);
         let mes_render = self.chat_template.apply_chat_template(&mes)?;
-
-        let input = self.pre_processor.process_info(&mes, &mes_render)?;
-        let mut input_ids = self
-            .tokenizer
-            .text_encode(input.replace_text.clone(), &self.device)?;
+        let (mes_text, pixel_values, image_grid_thw, pixel_values_video, video_grid_thw) =
+            if let Some(processor) = &self.pre_processor {
+                let input = processor.process_info(&mes, &mes_render)?;
+                (
+                    input.replace_text,
+                    input.pixel_values,
+                    input.image_grid_thw,
+                    input.pixel_values_video,
+                    input.video_grid_thw,
+                )
+            } else {
+                (mes_render, None, None, None, None)
+            };
+        // let input = self.pre_processor.process_info(&mes, &mes_render)?;
+        let mut input_ids = self.tokenizer.text_encode(mes_text, &self.device)?;
         let mut seq_len = input_ids.dim(1)?;
         let prompt_tokens = seq_len as u32;
         let mut seqlen_offset = 0;
-        let mut pixel_values = input.pixel_values.as_ref();
-        let image_grid_thw = input.image_grid_thw.as_ref();
-        let mut pixel_values_video = input.pixel_values_video.as_ref();
-        let video_grid_thw = input.video_grid_thw.as_ref();
+        let mut pixel_values = pixel_values.as_ref();
+        let image_grid_thw = image_grid_thw.as_ref();
+        let mut pixel_values_video = pixel_values_video.as_ref();
+        let video_grid_thw = video_grid_thw.as_ref();
         let mut generate = Vec::new();
         let sample_len = mes.max_tokens.unwrap_or(1024);
         for _ in 0..sample_len {
@@ -202,19 +212,30 @@ impl<'a> GenerateModel for Qwen3_5GenerateModel<'a> {
         let seed = mes.seed.unwrap_or(34562) as u64;
         let mut logit_processor = get_logit_processor(mes.temperature, mes.top_p, None, seed);
         let mes_render = self.chat_template.apply_chat_template(&mes)?;
-        let input = self.pre_processor.process_info(&mes, &mes_render)?;
-        let mut input_ids = self
-            .tokenizer
-            .text_encode(input.replace_text.clone(), &self.device)?;
+        // let input = self.pre_processor.process_info(&mes, &mes_render)?;
+        let (mes_text, pixel_values, image_grid_thw, pixel_values_video, video_grid_thw) =
+            if let Some(processor) = &self.pre_processor {
+                let input = processor.process_info(&mes, &mes_render)?;
+                (
+                    input.replace_text,
+                    input.pixel_values,
+                    input.image_grid_thw,
+                    input.pixel_values_video,
+                    input.video_grid_thw,
+                )
+            } else {
+                (mes_render, None, None, None, None)
+            };
+        let mut input_ids = self.tokenizer.text_encode(mes_text, &self.device)?;
         let mut seq_len = input_ids.dim(1)?;
         let mut seqlen_offset = 0;
         let sample_len = mes.max_tokens.unwrap_or(1024);
         let stream = stream! {
             let mut error_tokens = Vec::new();
-            let mut pixel_values = input.pixel_values.as_ref();
-            let image_grid_thw = input.image_grid_thw.as_ref();
-            let mut pixel_values_video = input.pixel_values_video.as_ref();
-            let video_grid_thw = input.video_grid_thw.as_ref();
+            let mut pixel_values = pixel_values.as_ref();
+            let image_grid_thw = image_grid_thw.as_ref();
+            let mut pixel_values_video = pixel_values_video.as_ref();
+            let video_grid_thw = video_grid_thw.as_ref();
             let mut tool_call_id = None;
             let mut tool_call_content = String::new();
             let mut generate = Vec::new();